+++ stories/SQFormScrollableCard.stories.js
@@ -11,11 +11,11 @@ export default {
   argTypes: {
     onSave: {action: 'onSave', table: {disable: true}},
     children: {table: {disable: true}},
-    validationSchema: {table: {disable: true}}
+    validationSchema: {table: {disable: true}},
   },
   parameters: {
-    docs: {page: createDocsPage({showStories: false})}
-  }
+    docs: {page: createDocsPage({showStories: false})},
+  },
 };
 
 const defaultArgs = {
@@ -23,17 +23,21 @@ const defaultArgs = {
   initialValues: {hello: ''},
   muiGridProps: {
     spacing: 2,
-    alignItems: 'center'
+    alignItems: 'center',
   },
   onSubmit: () => {},
-  validationSchema: {hello: Yup.string().required('Required')}
+  validationSchema: {hello: Yup.string().required('Required')},
 };
 
-const Template = args => {
+const Template = (args) => {
   const {wrapper, validationSchema, ...restArgs} = args;
 
   const basicCard = (
-    <SQFormScrollableCard {...restArgs}>
+    <SQFormScrollableCard
+      validationSchema={validationSchema}
+      shouldRequireFieldUpdates={true}
+      {...restArgs}
+    >
       <SQFormTextField
         name="hello"
         label="Hello"
@@ -140,7 +144,7 @@ const Template = args => {
 
 export const Default = Template.bind({});
 Default.args = {
-  ...defaultArgs
+  ...defaultArgs,
 };
 
 const SubHeader = () => (
@@ -152,7 +156,7 @@ const SubHeader = () => (
 export const WithSubHeader = Template.bind({});
 WithSubHeader.args = {
   ...defaultArgs,
-  SubHeaderComponent: <SubHeader />
+  SubHeaderComponent: <SubHeader />,
 };
 
 const Wrapper = ({children}) => (
@@ -173,7 +177,7 @@ WithSelfBoundingHeight.args = {
   title: 'With Self-Bounding Height',
   isSelfBounding: true,
   wrapper: <Wrapper />,
-  SubHeaderComponent: <SubHeader />
+  SubHeaderComponent: <SubHeader />,
 };
 
 export const WithStaticHeight = Template.bind({});
@@ -181,5 +185,5 @@ WithStaticHeight.args = {
   ...defaultArgs,
   title: 'With Static Height',
   height: 450,
-  SubHeaderComponent: <SubHeader />
+  SubHeaderComponent: <SubHeader />,
 };